Organize builder component stories in folder.
